social auth refactored. soucial auth endpoint doc fixed
