+++ users/serializers/__init__.py
@@ -16,6 +16,7 @@ from .artist_profile import (
 )
 from .artist_registration import ArtistRegistrationSerializer
 from .base_registration import BaseRegistrationSerializer
+from .jwt import LogoutSerializer, TokenPairSerializer
 from .listener_profile import ListenerMeSerializer
 from .listener_registration import ListenerRegistrationSerializer
 
@@ -36,4 +37,6 @@ __all__ = [
     'PhoneChangeSerializer',
     'EmptySerializer',
     'BecomeArtistSerializer',
+    'LogoutSerializer',
+    'TokenPairSerializer',
 ]
