+++ users/serializers/jwt.py
@@ -7,3 +7,10 @@ class LogoutSerializer(serializers.Serializer):
     """Сериализатор выхода пользователя."""
 
     refresh = serializers.CharField(label='Refresh токен')
+
+
+class TokenPairSerializer(serializers.Serializer):
+    """Пара JWT токенов."""
+
+    access = serializers.CharField()
+    refresh = serializers.CharField()
