+++ users/services/__init__.py
@@ -8,11 +8,6 @@ from .password_reset import (
     generate_password_reset_data,
     verify_password_reset_token,
 )
-from .social_auth import (
-    ensure_listener_profile,
-    issue_tokens_for_user,
-    login_with_social_data,
-)
 from .utils import (
     get_user_from_uid,
     set_user_password,
@@ -27,7 +22,4 @@ __all__ = [
     'get_user_from_uid',
     'set_user_password',
     'verify_email_token',
-    'login_with_social_data',
-    'issue_tokens_for_user',
-    'ensure_listener_profile',
 ]
